2.1-databases/work_with_database/requirements.txt
django
psycopg2-binary
python-dotenv
Pillow
gunicorn
flake8
djangorestframework